Extract key-equals-value file parsing out of misc_info.{cc,h}
Bug: b/438847412
