+++ base/cvd/cuttlefish/common/libs/key_equals_value/BUILD.bazel
@@ -1,0 +1,17 @@
+load("//cuttlefish/bazel:rules.bzl", "cf_cc_library")
+
+package(
+    default_visibility = ["//:android_cuttlefish"],
+)
+
+cf_cc_library(
+    name = "key_equals_value",
+    srcs = ["key_equals_value.cc"],
+    hdrs = ["key_equals_value.h"],
+    deps = [
+        "//cuttlefish/common/libs/fs",
+        "//cuttlefish/common/libs/utils:contains",
+        "//cuttlefish/common/libs/utils:result",
+        "//libbase",
+    ],
+)
